pyproject.toml
[project]
name = "g3ku-ai"
version = "1.0.0"
description = "A lightweight personal AI assistant framework"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "g3ku contributors"}
]
keywords = ["ai", "agent", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.20.0,<1.0.0",
    "litellm>=1.81.5,<2.0.0",
    "langgraph>=1.0.10,<2.0.0",
    "langgraph-checkpoint-sqlite>=3.0.0,<4.0.0",
    "aiosqlite>=0.21.0,<1.0.0",
    "langchain-core>=1.2.17,<2.0.0",
    "langchain>=1.2.10,<2.0.0",
    "langchain-openai>=1.1.10,<2.0.0",
    "langchain-qdrant>=1.1.0,<2.0.0",
    "langchain-anthropic>=1.3.4,<2.0.0",
    "langchain-google-genai>=4.2.1,<5.0.0",
    "langchain-community>=0.4.1,<1.0.0",
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "fastapi>=0.100.0,<1.0.0",
    "uvicorn>=0.20.0,<1.0.0",
    "websockets>=16.0,<17.0",
    "websocket-client>=1.9.0,<2.0.0",
    "httpx>=0.28.0,<1.0.0",
    "requests>=2.32.5,<3.0.0",
    "urllib3>=2.2.0,<3.0.0",
    "charset-normalizer>=3.4.0,<4.0.0",
    "chardet>=5.0.0,<6.0.0",
    "oauth-cli-kit>=0.1.3,<1.0.0",
    "loguru>=0.7.3,<1.0.0",
    "readability-lxml>=0.8.4,<1.0.0",
    "rich>=14.0.0,<15.0.0",
    "croniter>=6.0.0,<7.0.0",
    "dingtalk-stream>=0.24.0,<1.0.0",
    "python-telegram-bot[socks]>=22.0,<23.0",
    "lark-oapi>=1.5.0,<2.0.0",
    "socksio>=1.0.0,<2.0.0",
    "python-socketio>=5.16.0,<6.0.0",
    "msgpack>=1.1.0,<2.0.0",
    "slack-sdk>=3.39.0,<4.0.0",
    "slackify-markdown>=0.2.0,<1.0.0",
    "qq-botpy>=1.2.0,<2.0.0",
    "python-socks[asyncio]>=2.8.0,<3.0.0",
    "prompt-toolkit>=3.0.50,<4.0.0",
    "mcp>=1.26.0,<2.0.0",
    "playwright>=1.40.0,<2.0.0",
    "json-repair>=0.57.0,<1.0.0",
    "qdrant-client>=1.10.0,<2.0.0",
    "PyYAML>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
matrix = [
    "matrix-nio[e2e]>=0.25.2",
    "mistune>=3.0.0,<4.0.0",
    "nh3>=0.2.17,<1.0.0",
]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
g3ku-agent = "g3ku.cli.commands:app"
g3ku-web = "g3ku.web.main:run"
g3ku = "g3ku.g3ku_cli:main"

[build-system]
requires = ["hatchling", "editables>=0.5,<1.0.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["g3ku"]

[tool.hatch.build.targets.wheel.sources]
"g3ku" = "g3ku"

# Include non-Python files in skills and templates
[tool.hatch.build]
include = [
    "g3ku/**/*.py",
    "g3ku/web/frontend/**/*.html",
    "g3ku/web/frontend/**/*.css",
    "g3ku/web/frontend/**/*.js",
    "g3ku/templates/**/*.md",
    "g3ku/skills/**/*.md",
    "g3ku/skills/**/*.sh",
    "g3ku/capabilities/**/*.py",
    "g3ku/capabilities/**/*.yaml",
    "g3ku/capabilities/**/*.json",
    "g3ku/capabilities/**/*.md",
]

[tool.hatch.build.targets.sdist]
include = [
    "g3ku/",
    "bridge/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel.force-include]
"bridge" = "g3ku/bridge"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]



